# frontend/openapi-snapshot.json
{
  "components": {
    "schemas": {
...
      "AvatarUploadAcceptedResponse": {
        "description": "202 response — avatar upload accepted; processing happens in the background.",
        "properties": {
          "message": {
            "default": "Avatar upload accepted",
            "title": "Message",
            "type": "string"
          },
          "status": {
            "default": "accepted",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "AvatarUploadAcceptedResponse",
        "type": "object"
      },
      "AvatarUploadConfirmRequest": {
        "description": "Confirm upload after the client uploaded directly to S3.",
        "properties": {
          "staging_key": {
            "description": "Staging key from upload_url response",
            "title": "Staging Key",
            "type": "string"
          }
        },
        "required": [
          "staging_key"
        ],
        "title": "AvatarUploadConfirmRequest",
        "type": "object"
      },
      "AvatarUploadUrlResponse": {
        "description": "Response with presigned URL for avatar upload.",
        "properties": {
          "expires_in": {
            "default": 300,
            "description": "URL TTL in seconds",
            "title": "Expires In",
            "type": "integer"
          },
          "staging_key": {
            "description": "Staging key for confirm step",
            "title": "Staging Key",
            "type": "string"
          },
          "upload_url": {
            "description": "Presigned URL for direct upload to S3",
            "title": "Upload Url",
            "type": "string"
          }
        },
        "required": [
          "upload_url",
          "staging_key"
        ],
        "title": "AvatarUploadUrlResponse",
        "type": "object"
      },
...
      "FCMTokenUpdate": {
        "properties": {
          "fcm_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fcm Token"
          }
        },
        "title": "FCMTokenUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NotificationItemResponse": {
        "properties": {
          "action_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Url"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "booking_id": {
            "format": "uuid",
            "title": "Booking Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "other_party_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other Party Name"
          },
          "ride_destination": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ride Destination"
          },
          "ride_id": {
            "format": "uuid",
            "title": "Ride Id",
            "type": "string"
          },
          "ride_origin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ride Origin"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "created_at",
          "booking_id",
          "ride_id"
        ],
        "title": "NotificationItemResponse",
        "type": "object"
      },
...
      "PaginatedNotificationsResponse": {
        "description": "Cursor-paginated notifications for GET /users/me/notifications.",
        "properties": {
          "has_more": {
            "default": false,
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "default": 20,
            "title": "Limit",
            "type": "integer"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "title": "PaginatedNotificationsResponse",
        "type": "object"
      },
...
      "UserRead": {
        "properties": {
          "avatar_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Key"
          },
          "avatar_status": {
            "default": "none",
            "title": "Avatar Status",
            "type": "string"
          },
          "avatar_url_medium": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "400x400 — main profile image.",
            "readOnly": true,
            "title": "Avatar Url Medium"
          },
          "avatar_url_small": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "150x150 — chat lists, small avatars.",
            "readOnly": true,
            "title": "Avatar Url Small"
          },
          "email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "is_admin": {
            "default": false,
            "title": "Is Admin",
            "type": "boolean"
          },
          "is_verified": {
            "default": false,
            "title": "Is Verified",
            "type": "boolean"
          },
          "phone_number": {
            "title": "Phone Number",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "full_name",
          "phone_number",
          "avatar_url_small",
          "avatar_url_medium"
        ],
        "title": "UserRead",
        "type": "object"
      },
...
      "UserUpdate": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "new_password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Password"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "phone_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Number"
          }
        },
        "title": "UserUpdate",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "app__domain__users__schema__MessageResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "default": "success",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "LinkUp API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/users/fcm-token": {
      "patch": {
        "operationId": "update_fcm_token_api_v1_users_fcm_token_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FCMTokenUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__domain__users__schema__MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Fcm Token",
        "tags": [
          "Users",
          "Users"
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "description": "Return the authenticated user's profile (including sensitive fields like verification state).",
        "operationId": "get_my_profile_api_v1_users_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get My Profile",
        "tags": [
          "Users",
          "Users"
        ]
      },
      "put": {
        "description": "Update the authenticated user's profile.\nAccepts a UserUpdate object and passes it through to the service unchanged.",
        "operationId": "update_my_profile_api_v1_users_me_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update My Profile",
        "tags": [
          "Users",
          "Users"
        ]
      }
    },
    "/api/v1/users/me/avatar": {
      "delete": {
        "description": "Remove profile photo: deletes avatars/{user_id}/ from S3 and clears avatar_key in DB.",
        "operationId": "remove_my_avatar_api_v1_users_me_avatar_delete",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarUploadAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove My Avatar",
        "tags": [
          "Users",
          "Users"
        ]
      }
    },
    "/api/v1/users/me/avatar/confirm": {
      "post": {
        "description": "Confirm upload after the client uploaded directly to S3. Updates avatar_key in DB immediately and enqueues work.\nProcessing (resize to 3 sizes) runs in the background. staging_key must start with avatars/staging/{user_id}_.",
        "operationId": "confirm_avatar_upload_api_v1_users_me_avatar_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvatarUploadConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarUploadAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Confirm Avatar Upload",
        "tags": [
          "Users",
          "Users"
        ]
      }
    },
    "/api/v1/users/me/avatar/upload-url": {
      "get": {
        "description": "Return a presigned URL for direct upload to S3.\nFlow:\n1. Client calls this endpoint → receives presigned URL + staging_key\n2. Client uploads directly to S3 via the URL (5–8s)\n3. Client calls POST /me/avatar/confirm with staging_key\n4. Server enqueues work and returns 202 immediately (~1s)\n5. Worker processes in the background (finalize + DB update)",
        "operationId": "get_avatar_upload_url_api_v1_users_me_avatar_upload_url_get",
        "parameters": [
          {
            "in": "query",
            "name": "filename",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarUploadUrlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Avatar Upload Url",
        "tags": [
          "Users",
          "Users"
        ]
      }
    },
    "/api/v1/users/me/last-seen": {
      "patch": {
        "operationId": "update_last_seen_api_v1_users_me_last_seen_patch",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Last Seen",
        "tags": [
          "Users",
          "Users"
        ]
      }
    },
    "/api/v1/users/me/notifications": {
      "get": {
        "description": "Cursor-paginated notifications: driver pending joins + passenger booking updates.",
        "operationId": "get_my_notifications_api_v1_users_me_notifications_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedNotificationsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get My Notifications",
        "tags": [
          "Users",
          "Users"
        ]
      }
    },
    "/api/v1/users/me/test-push": {
      "post": {
        "operationId": "test_push_api_v1_users_me_test_push_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Test Push",
        "tags": [
          "Users",
          "Users"
        ]
      }
    },
    "/api/v1/users/{user_id}/last-seen": {
      "get": {
        "description": "Source for chat UI `last_seen`; called from chat-ws on GET /presence/{id}.",
        "operationId": "get_user_last_seen_api_v1_users__user_id__last_seen_get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get User Last Seen",
        "tags": [
          "Users",
          "Users"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "Local",
      "url": "http://127.0.0.1:8000"
    }
  ]
}